refactoring, SSL support added

+++ src/main/java/com/purple/ua/universityappointment/dto/AppointmentDto.java
@@ -21,13 +21,13 @@ public class AppointmentDto {
     private Status status;
 
     @NotNull
-    @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime fromDate;
 
     @NotNull
-    @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime toDate;
 
-    private int finalPrice;
+    private double finalPrice;
 
 }
